requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
uuid6==1.0.3
